[openapi-gen] Move Endpoint to be shared by all generation
