+++ src/it/scala/temple/generate/service/GoGeneratorIntegrationTestData.scala
@@ -1,7 +1,6 @@
-package temple.generate.language.service
+package temple.generate.service
 
-import temple.generate.language.service.adt._
-import temple.utils.FileUtils._
+import temple.generate.Endpoint
 
 object GoGeneratorIntegrationTestData {
