+++ src/main/scala/temple/generate/service/ServiceRoot.scala
@@ -1,3 +1,5 @@
-package temple.generate.language.service.adt
+package temple.generate.service
+
+import temple.generate.Endpoint
 
 case class ServiceRoot(name: String, module: String, comms: Seq[String], endpoints: Set[Endpoint], port: Int)
