+++ src/main/scala/temple/generate/service/go/GoServiceGenerator.scala
@@ -1,7 +1,7 @@
-package temple.generate.language.service.go
+package temple.generate.service.go
 
-import temple.generate.language.service.ServiceGenerator
-import temple.generate.language.service.adt._
+import temple.generate.Endpoint
+import temple.generate.service.{ServiceGenerator, ServiceRoot}
 import temple.generate.utils.CodeTerm
 import temple.generate.utils.CodeTerm.{CodeWrap, mkCode}
 import temple.utils.FileUtils
@@ -127,7 +127,7 @@ object GoServiceGenerator extends ServiceGenerator {
     FileUtils.readResources("go/genFiles/json_middleware.go").stripLineEnd
 
   private def generateHandler(serviceName: String, endpoint: Endpoint): String =
-    s"func $serviceName${endpoint.verb}Handler(w http.ResponseWriter, r *http.Request) {}"
+    s"func $serviceName${verb(endpoint)}Handler(w http.ResponseWriter, r *http.Request) {}"
 
   private def generateHandlers(serviceName: String, endpoints: Set[Endpoint]): String =
     mkCode.doubleLines(
@@ -211,5 +211,13 @@ object GoServiceGenerator extends ServiceGenerator {
         generateCommInit(),
       ),
     )).map { case (path, contents) => path -> (contents + "\n") }
+  }
+
+  private def verb(endpoint: Endpoint): String = endpoint match {
+    case Endpoint.ReadAll => "List"
+    case Endpoint.Create  => "Create"
+    case Endpoint.Read    => "Read"
+    case Endpoint.Update  => "Update"
+    case Endpoint.Delete  => "Delete"
   }
 }
